requirements.txt
streamlit
pandas
langchain
langchain-openai
python-decouple
openpyxl
langchain langchain-community